UserManagement/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-authtoken==3.15.2
drf-yasg==1.21.7
Pillow==10.0.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0